+++ src/config-type.ts
@@ -38,6 +38,14 @@ export interface CommonConfig {
    */
   cors: boolean | HeadersProvider;
   /**
+   * @desc How to respond to a request that uses a wrong method to an existing endpoint
+   * @example 404 — Not found
+   * @example 405 — Method not allowed, incl. the "Allowed" header with a list of methods
+   * @default 404
+   * @todo consider changing default to 405 in v23
+   * */
+  wrongMethodBehavior?: 404 | 405;
+  /**
    * @desc The ResultHandler to use for handling routing, parsing and upload errors
    * @default defaultResultHandler
    * @see defaultResultHandler